+++ src/main/java/com/ict/careus/model/campaign/Category.java
@@ -14,7 +14,7 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "category_id")
-    private int categoryId;
+    private Long categoryId;
 
     @Enumerated(EnumType.STRING)
     private CampaignCategory categoryName;
